[Refactoring] Split EventSender into submodules
